refactor(theme): move theme into theme/
