+++ src/styled-components.ts
@@ -1,13 +1,13 @@
 import * as styledComponents from 'styled-components';
 import { ThemedStyledComponentsModule } from 'styled-components';
-import { ThemeInterface } from './theme';
+import { ThemeInterface } from './theme/theme-type';
 
 const {
   default: styled,
   css,
   createGlobalStyle,
   keyframes,
-  ThemeProvider
+  ThemeProvider,
 } = styledComponents as ThemedStyledComponentsModule<ThemeInterface>;
 
 export { styled, css, createGlobalStyle, keyframes, ThemeProvider };
